# Convert `SmallSet` to be a bitset