--- a/src/full/Agda/Utils/SmallSet.hs
+++ b/src/full/Agda/Utils/SmallSet.hs
@@ -1,8 +1,6 @@
--- | Small sets represented as immutable bit arrays for fast membership checking.
+-- | Small sets represented as a bitmask for fast membership checking.
 --
--- Membership checking is O(1), but all other operations are O(n)
--- where n is the size of the element type.
--- The element type needs to implement 'Bounded' and 'Ix'.
+-- With the exception of converting to/from lists, all operations are O(1).
 --
 -- Mimics the interface of 'Data.Set'.
 --
@@ -14,9 +12,9 @@
 
 module Agda.Utils.SmallSet
   ( SmallSet()
+  , SmallSetElement
   , Ix
   , (\\)
-  , complement
   , delete
   , difference
   , elems
@@ -24,8 +22,6 @@ module Agda.Utils.SmallSet
   , fromList, fromAscList, fromDistinctAscList
   , insert
   , intersection
-  , isSubsetOf
-  , mapMembership
   , member
   , notMember
   , null
@@ -33,126 +29,96 @@ module Agda.Utils.SmallSet
   , toList, toAscList
   , total
   , union
-  , zipMembershipWith
   ) where
 
 import Prelude hiding (null)
 
 import Control.DeepSeq
 
-import Data.Array.IArray (Ix, Array)
-import qualified Data.Array.IArray as Array
+import Data.Word (Word64)
+import Data.List (foldl')
+import Data.Bits
+import Data.Ix
 
--- Note: we might want to use unboxed arrays, but they have no Data instance
+-- | An element in a small set.
 --
--- Update: There is currently no need for a Data instance. An attempt
--- was made to replace Array with Data.Array.Unboxed.UArray. Limited
--- testing suggested that this does not make much of a difference in
--- practice (at least not when it comes to type-checking the standard
--- library up to and including Data.Nat, with Agda compiled without
--- -foptimise-heavily).
+-- This must implement 'Bounded' and 'Ix', and contain at most 64 values.
+class (Bounded a, Ix a) => SmallSetElement a where
 
--- | Let @n@ be the size of type @a@.
-type SmallSetElement a = (Bounded a, Ix a)
-newtype SmallSet a = SmallSet { theSmallSet :: Array a Bool }
+newtype SmallSet a = SmallSet { theSmallSet :: Word64 }
   deriving (Eq, Ord, Show, NFData)
 
 -- * Query
 
--- | Time O(n)!
+-- | Time O(1).
 null :: SmallSetElement a => SmallSet a -> Bool
-null = all (== False) . Array.elems . theSmallSet
+null s = theSmallSet s == 0
 
 -- | Time O(1).
 member :: SmallSetElement a => a -> SmallSet a -> Bool
-member a s = theSmallSet s Array.! a
+member a s = theSmallSet s `testBit` idx a
 
 -- | @not . member a@.  Time O(1).
 notMember :: SmallSetElement a => a -> SmallSet a -> Bool
 notMember a = not . member a
 
--- | Time O(n).
-isSubsetOf ::  SmallSetElement a => SmallSet a -> SmallSet a -> Bool
-isSubsetOf s t = and $ toBoolListZipWith implies s t
-  where implies a b = if a then b else True
-
 -- * Construction
 
--- | The empty set.  Time O(n).
+-- | The empty set.  Time O(1).
 empty :: SmallSetElement a => SmallSet a
-empty = fromBoolList (repeat False)
+empty = SmallSet 0
 
--- | The full set.  Time O(n).
-total :: SmallSetElement a => SmallSet a
-total = fromBoolList (repeat True)
+-- | The full set.  Time O(1).
+total :: forall a. SmallSetElement a => SmallSet a
+total = SmallSet mask where
+  size = rangeSize (bounds :: (a, a))
+  mask = (1 `shiftL` size) - 1
 
--- | A singleton set.  Time O(n).
+-- | A singleton set.  Time O(1).
 singleton :: SmallSetElement a => a -> SmallSet a
-singleton a = fromList [a]
+singleton a = SmallSet $ bit (idx a)
 
--- | Time O(n).
+-- | Time O(1).
 insert :: SmallSetElement a => a -> SmallSet a -> SmallSet a
-insert a = update [(a,True)]
+insert a s = SmallSet $ theSmallSet s `setBit` idx a
 
--- | Time O(n).
+-- | Time O(1).
 delete :: SmallSetElement a => a -> SmallSet a -> SmallSet a
-delete a = update [(a,False)]
+delete a s = SmallSet $ theSmallSet s `clearBit` idx a
 
 -- * Combine
 
--- | Time O(n).
-complement :: SmallSetElement a => SmallSet a -> SmallSet a
-complement = mapMembership not
-
--- | Time O(n).
+-- | Time O(1).
 difference, (\\) :: SmallSetElement a => SmallSet a -> SmallSet a -> SmallSet a
-difference = zipMembershipWith $ \ b c -> b && not c
+difference s t = SmallSet $ theSmallSet s .&. complement (theSmallSet t)
 (\\)       = difference
 
--- | Time O(n).
+-- | Time O(1).
 intersection ::  SmallSetElement a => SmallSet a -> SmallSet a -> SmallSet a
-intersection = zipMembershipWith (&&)
+intersection s t = SmallSet $ theSmallSet s .&. theSmallSet t
 
 -- | Time O(n).
 union ::  SmallSetElement a => SmallSet a -> SmallSet a -> SmallSet a
-union = zipMembershipWith (||)
-
--- | Time O(n).
-mapMembership :: SmallSetElement a => (Bool -> Bool) -> SmallSet a -> SmallSet a
-mapMembership f = SmallSet . Array.amap f . theSmallSet
-
--- | Time O(n).
-zipMembershipWith :: SmallSetElement a => (Bool -> Bool -> Bool) -> SmallSet a -> SmallSet a -> SmallSet a
-zipMembershipWith f s t = fromBoolList $ toBoolListZipWith f s t
+union s t = SmallSet $ theSmallSet s .|. theSmallSet t
 
 -- * Conversion
 
 -- | Time O(n).
 elems, toList, toAscList :: SmallSetElement a => SmallSet a -> [a]
-elems     = map fst . filter snd . Array.assocs . theSmallSet
+elems s   = filter (\i -> theSmallSet s `testBit` idx i) (range bounds)
 toList    = elems
 toAscList = elems
 
 -- | Time O(n).
 fromList, fromAscList, fromDistinctAscList :: SmallSetElement a => [a] -> SmallSet a
-fromList            = flip update empty . map (,True)
+fromList            = foldl' (flip insert) empty
 fromAscList         = fromList
 fromDistinctAscList = fromList
 
--- * Internal
+-- * Internals
 
--- | Time O(n).  Assumes @Bool@-vector of length @n@.
-fromBoolList :: SmallSetElement a => [Bool] -> SmallSet a
-fromBoolList = SmallSet . Array.listArray (minBound, maxBound)
+bounds :: SmallSetElement a => (a, a)
+bounds = (minBound, maxBound)
 
--- | Time O(n).  Produces @Bool@-vector of length @n@.
-toBoolList :: SmallSetElement a => SmallSet a -> [Bool]
-toBoolList = Array.elems . theSmallSet
-
--- | Time O(n).  Produces @Bool@-vector of length @n@.
-toBoolListZipWith :: SmallSetElement a => (Bool -> Bool -> Bool) -> SmallSet a -> SmallSet a -> [Bool]
-toBoolListZipWith f s t = zipWith f (toBoolList s) (toBoolList t)
-
--- | Time O(n).  Bulk insert/delete.
-update :: SmallSetElement a => [(a,Bool)] -> SmallSet a -> SmallSet a
-update u s = SmallSet $ theSmallSet s Array.// u
+idx :: SmallSetElement a => a -> Int
+idx a = index bounds a