This was discussed when originally implemented in bda742e7, and after some profiling it feels like it's worth it, reducing time taken in type-checking by ~2%.

This does have the limitation that types with at most 64 values can be used as set elements. I don't think that's an issue in practice - the previous implementation would have been expensive with this many elements - but have changed it so you need to implement `SmallSetElement` explicitly, to make this requirement more clear.

This also drops a couple of unused SmallSet methods (`complement`, `isSubsetOf`) which didn't feel worth implementing. Happy to add those back if preferred.

<details><summary>Performance measurements</summary>

Numbers were taken using GHC 9.2.7 and running Agda against `src/Data/Nat.lagda.md` from  https://github.com/plt-amy/1lab/commit/3b51679d95520c15ccd20accb9ab686e1eaa3c55. Agda was run under `perf stat` and with `+RTS -A128M -s -RTS --profile=internal`

### Master (791fb5e7b327d5e9b15146e5c234cfdbb6c5c14d)

```
Total                        32,597ms           
Miscellaneous                   122ms           
Typing                        1,317ms (22,605ms)
Typing.CheckRHS              10,147ms           
Typing.OccursCheck            5,761ms           
Typing.CheckLHS               1,504ms  (1,696ms)
Typing.CheckLHS.UnifyIndices    192ms           
Typing.TypeSig                1,594ms           
Typing.InstanceSearch         1,315ms           
Typing.Generalize               710ms           
Typing.With                      61ms           
Parsing                         385ms  (2,869ms)
Parsing.OperatorsExpr         1,983ms           
Parsing.OperatorsPattern        499ms           
Serialization                 1,530ms  (1,972ms)
Serialization.Compress          250ms           
Serialization.BinaryEncode      131ms           
Serialization.Sort               37ms           
Serialization.BuildInterface     22ms           
Scoping                         518ms  (1,530ms)
Scoping.InverseScopeLookup    1,011ms           
Deserialization                 671ms    (943ms)
Deserialization.Compaction      271ms           
Coverage                        925ms           
Highlighting                    548ms           
Positivity                      389ms           
DeadCode                        346ms           
Termination                     115ms    (217ms)
Termination.RecCheck             71ms           
Termination.Compare              16ms           
Termination.Graph                14ms           
Injectivity                      68ms           
Import                           58ms           
  60,075,994,392 bytes allocated in the heap
   2,025,767,248 bytes copied during GC
     102,845,184 bytes maximum residency (24 sample(s))
         509,128 bytes maximum slop
             343 MiB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0       429 colls,     0 par    2.940s   2.943s     0.0069s    0.0422s
  Gen  1        24 colls,     0 par    0.663s   0.663s     0.0276s    0.0835s

  TASKS: 4 (1 bound, 3 peak workers (3 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.004s  (  0.004s elapsed)
  MUT     time   29.011s  ( 28.905s elapsed)
  GC      time    3.602s  (  3.606s elapsed)
  EXIT    time    0.001s  (  0.006s elapsed)
  Total   time   32.619s  ( 32.521s elapsed)

  Alloc rate    2,070,797,507 bytes per MUT second

  Productivity  88.9% of total user, 88.9% of total elapsed


 Performance counter stats for 'scratch/agda-791fb5e7b327d5e9b15146e5c234cfdbb6c5c14d src/Data/Nat.lagda.md +RTS -A128M -s -RTS --profile=internal':

         32,612.34 msec task-clock:u                     #    1.002 CPUs utilized             
                 0      context-switches:u               #    0.000 /sec                      
                 0      cpu-migrations:u                 #    0.000 /sec                      
           108,090      page-faults:u                    #    3.314 K/sec                     
   132,493,868,462      cycles:u                         #    4.063 GHz                         (83.09%)
     8,691,989,623      stalled-cycles-frontend:u        #    6.56% frontend cycles idle        (83.39%)
    26,582,563,560      stalled-cycles-backend:u         #   20.06% backend cycles idle         (83.36%)
   134,893,320,440      instructions:u                   #    1.02  insn per cycle            
                                                  #    0.20  stalled cycles per insn     (83.38%)
    25,729,291,857      branches:u                       #  788.943 M/sec                       (83.40%)
     1,003,666,757      branch-misses:u                  #    3.90% of all branches             (83.39%)

      32.544316751 seconds time elapsed

      32.137447000 seconds user
       0.505074000 seconds sys
```

### This PR
```
Total                        31,964ms           
Miscellaneous                   128ms           
Typing                        1,263ms (21,818ms)
Typing.CheckRHS               9,785ms           
Typing.OccursCheck            5,420ms           
Typing.CheckLHS               1,504ms  (1,686ms)
Typing.CheckLHS.UnifyIndices    181ms           
Typing.TypeSig                1,591ms           
Typing.InstanceSearch         1,311ms           
Typing.Generalize               713ms           
Typing.With                      45ms           
Parsing                         418ms  (2,904ms)
Parsing.OperatorsExpr         2,011ms           
Parsing.OperatorsPattern        474ms           
Serialization                 1,681ms  (2,147ms)
Serialization.Compress          251ms           
Serialization.BinaryEncode      145ms           
Serialization.Sort               47ms           
Serialization.BuildInterface     20ms           
Scoping                         513ms  (1,495ms)
Scoping.InverseScopeLookup      982ms           
Deserialization                 617ms    (897ms)
Deserialization.Compaction      280ms           
Coverage                        894ms           
Highlighting                    549ms           
DeadCode                        406ms           
Positivity                      395ms           
Termination                      90ms    (203ms)
Termination.RecCheck             71ms           
Termination.Compare              26ms           
Termination.Graph                14ms           
Injectivity                      66ms           
Import                           57ms           
  56,391,638,208 bytes allocated in the heap
   2,115,290,920 bytes copied during GC
     121,565,144 bytes maximum residency (21 sample(s))
         495,656 bytes maximum slop
             396 MiB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0       405 colls,     0 par    2.955s   2.958s     0.0073s    0.0382s
  Gen  1        21 colls,     0 par    0.782s   0.782s     0.0373s    0.1090s

  TASKS: 4 (1 bound, 3 peak workers (3 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.004s  (  0.004s elapsed)
  MUT     time   28.243s  ( 28.111s elapsed)
  GC      time    3.737s  (  3.740s elapsed)
  EXIT    time    0.001s  (  0.005s elapsed)
  Total   time   31.986s  ( 31.860s elapsed)

  Alloc rate    1,996,627,860 bytes per MUT second

  Productivity  88.3% of total user, 88.2% of total elapsed


 Performance counter stats for 'scratch/agda-c45cdd75d95b05f89dc42eceda96bc129b5f2e01 src/Data/Nat.lagda.md +RTS -A128M -s -RTS --profile=internal':

         31,965.14 msec task-clock:u                     #    1.003 CPUs utilized             
                 0      context-switches:u               #    0.000 /sec                      
                 0      cpu-migrations:u                 #    0.000 /sec                      
           120,868      page-faults:u                    #    3.781 K/sec                     
   129,293,504,726      cycles:u                         #    4.045 GHz                         (83.31%)
     7,465,832,160      stalled-cycles-frontend:u        #    5.77% frontend cycles idle        (83.31%)
    28,042,847,329      stalled-cycles-backend:u         #   21.69% backend cycles idle         (83.34%)
   131,288,667,216      instructions:u                   #    1.02  insn per cycle            
                                                  #    0.21  stalled cycles per insn     (83.33%)
    25,027,260,626      branches:u                       #  782.955 M/sec                       (83.32%)
       982,319,879      branch-misses:u                  #    3.92% of all branches             (83.38%)

      31.885410422 seconds time elapsed

      31.475750000 seconds user
       0.535415000 seconds sys
```

</details>