@SquidDev 2023-04-28
> It would also set non-existent elements to 1, but maybe this wouldn't be a problem?

Yeah, I avoided that originally as it felt a little undisciplined, but I don't think it's an issue in practice. Will make that change :).

>  Surprised that this should make such difference

Likewise! I've not seen it come up in other profiles people have done (such as those in #6090), but it definitely has an impact. Most of this is due to the fact that `SmallSet.\\` (which is quite allocation happy) is called from `shouldTryFastReduce` (which is a relatively hot function):

https://github.com/agda/agda/blob/791fb5e7b327d5e9b15146e5c234cfdbb6c5c14d/src/full/Agda/TypeChecking/Reduce.hs#L527

I think that expression could be rewritten to be ``requiredReductions `isSubSet` allowed && allowed `isSubSet` allReductions``, which might behave a little better, but changing the underlying implementation seemed the right route.